fixed some small errors

+++ Engine/Source/Core/Input/InputEvents.h
@@ -1,13 +1,13 @@
 #pragma once
 struct MouseButtonEvent
 {
-    MouseButtonEvent(int x, int y, int button, bool down)
+    MouseButtonEvent(double x, double y, int button, bool down)
         : x(x), y(y), button(button), down(down)
     {
     }
 
-    int x;
-    int y;
+    double x;
+    double y;
     int button;
     bool down;
 };
